+++ src/test/java/io/github/astrapi69/file/create/FileFactoryTest.java
@@ -34,8 +34,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-import io.github.astrapi69.file.create.model.FileCreationState;
-import io.github.astrapi69.file.create.model.FileInfo;
 import org.apache.commons.lang3.SystemUtils;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -44,6 +42,8 @@ import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
 import io.github.astrapi69.file.FileTestCase;
+import io.github.astrapi69.file.create.model.FileCreationState;
+import io.github.astrapi69.file.create.model.FileInfo;
 import io.github.astrapi69.file.delete.DeleteFileExtensions;
 import io.github.astrapi69.file.system.SystemFileExtensions;